+++ doc/changelog.md
@@ -14,6 +14,7 @@ Jump to:
 Description
 
 - Add helper methods to MLI classes
+- Update error handling for consistency
 - Parameterize installation of dragon package with `smart build`
 - Update docstrings
 - Implement asynchronous notifications for shared data